Render scenario, session and review content from the layout state
