+++ support/lobby/lib/Sessions.js
@@ -1,15 +1,13 @@
 import React from "react";
 import { Table, Button } from "react-bootstrap";
 
-import * as locals from "./locals";
-
 export default function Sessions( props ) {
   return <Table striped>
     <thead>
       <Head/>
     </thead>
     <tbody>
-      <Sessioons instructor={ props.instructor }/>
+      <Sessioons records={ props.records } instructor={ props.instructor }/>
     </tbody>
   </Table>;
 }
@@ -33,10 +31,8 @@ function Head( props ) {
 }
 
 function Sessioons( props ) {
-  const records =
-    locals.scenarioScenarioSessions( locals.manifest[ "/ITDG/index.vwf" ] || [] );
   return <React.Fragment>
-    { records.map( ( record, index ) => <Session key={ index } { ...record } instructor={ props.instructor }/> ) }
+    { props.records.map( ( record, index ) => <Session key={ index } { ...record } instructor={ props.instructor }/> ) }
   </React.Fragment>;
 }
 
@@ -48,7 +44,7 @@ function Session( props ) {
       <td>
         { session.state.scenarioTitle }
         <br/>
-        <span className="small">{ locals.instructorStudentsLabel( session ) }</span>
+        <span className="small">{ instructorStudentsLabel( session ) }</span>
       </td><td>
         { session.state.classroom.company }
       </td><td>
@@ -66,4 +62,29 @@ function Session( props ) {
   } else {
     return null;
   }
+}
+
+// Generate the Instructor/Students annotation for a session.
+
+export function instructorStudentsLabel( session ) {
+
+  var instanceCounts = session.completion.instance || { instructors: 0, students: 0 },
+    label = "";
+
+  if ( instanceCounts.instructors > 0 || instanceCounts.students > 0 ) {
+
+    if ( instanceCounts.instructors > 0 ) {
+      label += "Instructor, ";
+    }
+
+    if ( instanceCounts.students === 1 ) {
+      label += instanceCounts.students + " student";
+    } else {
+      label += instanceCounts.students + " students";
+    }
+
+  }
+
+  return label;
+
 }
